Add API and model layer capability for link (URL) files

+++ girder/models/file.py
@@ -40,27 +40,74 @@ class File(Model):
         file is stored in, and call deleteFile on it, then delete the file
         record from the database.
         """
-        assetstore = self.model('assetstore').load(file['assetstoreId'])
-        adapter = assetstore_utilities.getAssetstoreAdapter(assetstore)
-        adapter.deleteFile(file)
+        if 'assetstoreId' in file:
+            assetstore = self.model('assetstore').load(file['assetstoreId'])
+            adapter = assetstore_utilities.getAssetstoreAdapter(assetstore)
+            adapter.deleteFile(file)
         Model.remove(self, file)
 
     def download(self, file, offset=0, headers=True):
         """
         Use the appropriate assetstore adapter for whatever assetstore the
-        file is stored in, and call downloadFile on it.
+        file is stored in, and call downloadFile on it. If the file is a link
+        file rather than a file in an assetstore, we redirect to it.
         """
-        assetstore = self.model('assetstore').load(file['assetstoreId'])
-        adapter = assetstore_utilities.getAssetstoreAdapter(assetstore)
-        return adapter.downloadFile(file, offset=offset, headers=headers)
+        if 'assetstoreId' in file:
+            assetstore = self.model('assetstore').load(file['assetstoreId'])
+            adapter = assetstore_utilities.getAssetstoreAdapter(assetstore)
+            return adapter.downloadFile(file, offset=offset, headers=headers)
+        elif 'linkUrl' in file:
+            raise cherrypy.HTTPRedirect(file['linkUrl'], status=302)
+        else:  # pragma: no cover
+            raise Exception('File has no known download mechanism.')
 
     def validate(self, doc):
+        if doc.get('assetstoreId') is None:
+            if 'linkUrl' not in doc:
+                raise ValidationException(
+                    'File must have either an assetstore ID or a link URL.',
+                    'linkUrl')
+            if not doc['linkUrl'].startswith(('http:', 'https:')):
+                raise ValidationException(
+                    'Linked file URL must start with http: or https:.',
+                    'linkUrl')
         if 'name' not in doc or not doc['name']:
             raise ValidationException('File name must not be empty.', 'name')
 
         doc['exts'] = doc['name'].split('.')[1:]
 
         return doc
+
+    def createLinkFile(self, name, parent, parentType, url, creator):
+        """
+        Create a file that is a link to a URL rather than something we maintain
+        in an assetstore.
+        :param name: The local name for the file.
+        :type name: str
+        :param parent: The parent object for this file.
+        :type parent: folder or item
+        :param parentType: The parent type (folder or item)
+        :type parentType: str
+        :param url: The URL that this file points to
+        :param creator: The user creating the file.
+        :type user: user
+        """
+        if parentType == 'folder':
+            # Create a new item with the name of the file.
+            item = self.model('item').createItem(
+                name=name, creator=user, folder=parent)
+        elif parentType == 'item':
+            item = parent
+
+        file = {
+            'created': datetime.datetime.now(),
+            'itemId': item['_id'],
+            'creatorId': creator['_id'],
+            'assetstoreId': None,
+            'name': name,
+            'linkUrl': url
+        }
+        return self.save(file)
 
     def createFile(self, creator, item, name, size, assetstore, mimeType):
         """
